+++ src/utils/functions/economy/levelling.ts
@@ -9,8 +9,8 @@ import { addKarma } from "../karma/karma";
 import { getUserId, MemberResolvable } from "../member";
 import { pluralize } from "../string";
 import { addNotificationToQueue, getDmSettings } from "../users/notifications";
-import { getLastKnownAvatar } from "../users/tag";
 import { addTag } from "../users/tags";
+import { getLastKnownAvatar } from "../users/username";
 import { addBalance, getBankBalance, removeBankBalance } from "./balance";
 import { addBooster, getBoosters } from "./boosters";
 import { addInventoryItem } from "./inventory";
