+++ src/utils/functions/economy/top.ts
@@ -8,8 +8,8 @@ import { Mutex } from "../mutex";
 import PageManager from "../page";
 import { escapeFormattingCharacters, formatTime, pluralize } from "../string";
 import { getPreferences } from "../users/notifications";
-import { getLastKnownUsername, updateLastKnownUsername } from "../users/tag";
 import { getActiveTag } from "../users/tags";
+import { getLastKnownUsername, updateLastKnownUsername } from "../users/username";
 import { checkLeaderboardPositions } from "./stats";
 import { getAchievements, getItems, getTagsData } from "./utils";
 import pAll = require("p-all");
